Move viewer specific modules under src/viewer directory
This helps reduce the mental overhead of all files residing under the
same directory level in src/
